chore: add options
