+++ test/mocks/user_transformation/test-data/transformation-mocks.ts
@@ -237,153 +237,3 @@ export const transformationMocks: Record<string, any> = {
     imports: [],
   },
 };
-
-// Mock library codes
-export const libraryMocks: Record<string, any> = {
-  'lodash-lib-v1': {
-    versionId: 'lodash-lib-v1',
-    code: `// Lodash library mock - simplified version
-      const _ = {
-        omit: (obj, keys) => {
-          const result = { ...obj };
-          keys.forEach(key => delete result[key]);
-          return result;
-        },
-        mapKeys: (obj, iteratee) => {
-          const result = {};
-          Object.keys(obj).forEach(key => {
-            result[iteratee(obj[key], key)] = obj[key];
-          });
-          return result;
-        },
-        camelCase: (str) => {
-          return str.replace(/[-_\\s]+(.)?/g, (_, c) => c ? c.toUpperCase() : '');
-        },
-        size: (obj) => {
-          return Object.keys(obj || {}).length;
-        }
-      };
-      export default _;
-    `,
-    name: 'lodash',
-    importName: 'lodash',
-  },
-
-  'moment-lib-v1': {
-    versionId: 'moment-lib-v1',
-    code: `// Moment library mock - simplified version
-      const moment = {
-        now: () => Date.now(),
-        format: (date, format) => new Date(date).toISOString(),
-        add: (date, amount, unit) => new Date(date.getTime() + amount * 1000),
-        subtract: (date, amount, unit) => new Date(date.getTime() - amount * 1000)
-      };
-      export default moment;
-    `,
-    name: 'moment',
-    importName: 'moment',
-  },
-};
-
-// Mock Rudder library codes
-export const rudderLibraryMocks: Record<string, any> = {
-  'urlParser@v1': {
-    code: `// URL Parser library mock
-      const URLSearchParams = class {
-        constructor(url) {
-          this.params = new Map();
-          if (url && url.includes('?')) {
-            const queryString = url.split('?')[1];
-            queryString.split('&').forEach(param => {
-              const [key, value] = param.split('=');
-              this.params.set(decodeURIComponent(key), decodeURIComponent(value || ''));
-            });
-          }
-        }
-        
-        get(key) {
-          return this.params.get(key);
-        }
-        
-        has(key) {
-          return this.params.has(key);
-        }
-      };
-      
-      export default { URLSearchParams };
-    `,
-    name: 'urlParser',
-    importName: '@rs/urlParser/v1',
-  },
-
-  'crypto@v1': {
-    code: `// Crypto library mock
-      const crypto = {
-        randomUUID: () => 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
-          const r = Math.random() * 16 | 0;
-          const v = c == 'x' ? r : (r & 0x3 | 0x8);
-          return v.toString(16);
-        }),
-        
-        createHash: (algorithm) => ({
-          update: (data) => ({
-            digest: (encoding) => {
-              // Simple hash mock
-              let hash = 0;
-              for (let i = 0; i < data.length; i++) {
-                const char = data.charCodeAt(i);
-                hash = ((hash << 5) - hash) + char;
-                hash = hash & hash;
-              }
-              return Math.abs(hash).toString(16);
-            }
-          })
-        })
-      };
-      
-      export default crypto;
-    `,
-    name: 'crypto',
-    importName: '@rs/crypto/v1',
-  },
-};
-
-// Mock external API responses for fetch calls made by user transformations
-export const externalApiMocks: Record<string, any> = {
-  'https://api.example.com/enrich': {
-    method: 'POST',
-    response: {
-      status: 200,
-      headers: { 'content-type': 'application/json' },
-      body: {
-        enriched: true,
-        timestamp: '2023-01-01T00:00:00Z',
-        source: 'external-api',
-      },
-    },
-  },
-
-  'https://api.example.com/user-profile': {
-    method: 'GET',
-    response: {
-      status: 200,
-      headers: { 'content-type': 'application/json' },
-      body: {
-        profile: {
-          segment: 'premium',
-          preferences: ['email', 'sms'],
-          lastSeen: '2023-01-01T00:00:00Z',
-        },
-      },
-    },
-  },
-
-  'https://api.example.com/error': {
-    method: 'GET',
-    response: {
-      status: 500,
-      headers: { 'content-type': 'application/json' },
-      body: { error: 'Internal server error' },
-    },
-  },
-};
